third_party/blink/web_tests/platform/generic/media/media-fragments/TC0086-expected.txt


Title: TC0086
Fragment: 'foo&t=3,7'
Comment: Rubbish before & is ignored.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
